+++ src/binary.rs
@@ -1,10 +1,15 @@
-﻿//! Capture 3. Binary Encoding
+//! Capture 3. Binary Encoding
 
-/// SBI functions return type.
+/// SBI functions return type
 ///
 /// > SBI functions must return a pair of values in a0 and a1,
 /// > with a0 returning an error code.
 /// > This is analogous to returning the C structure `SbiRet`.
+///
+/// Note: if this structure is used in function return on conventional
+/// Rust code, it would not require to pin memory representation as
+/// extern C. The `repr(C)` is set in case that some users want to use
+/// this structure in FFI code.
 #[derive(Clone, Copy, PartialEq, Eq)]
 #[repr(C)]
 pub struct SbiRet {
@@ -14,14 +19,23 @@ pub struct SbiRet {
     pub value: usize,
 }
 
+/// SBI success state return value
 pub const RET_SUCCESS: usize = 0;
+/// Error for SBI call failed for unknown reasons
 pub const RET_ERR_FAILED: usize = -1isize as _;
+/// Error for target operation not supported
 pub const RET_ERR_NOT_SUPPORTED: usize = -2isize as _;
+/// Error for invalid parameter
 pub const RET_ERR_INVALID_PARAM: usize = -3isize as _;
+/// Error for denied (unused in standard extensions)
 pub const RET_ERR_DENIED: usize = -4isize as _;
+/// Error for invalid address
 pub const RET_ERR_INVALID_ADDRESS: usize = -5isize as _;
+/// Error for resource already available
 pub const RET_ERR_ALREADY_AVAILABLE: usize = -6isize as _;
+/// Error for resource already started
 pub const RET_ERR_ALREADY_STARTED: usize = -7isize as _;
+/// Error for resource already stopped
 pub const RET_ERR_ALREADY_STOPPED: usize = -8isize as _;
 
 impl core::fmt::Debug for SbiRet {
@@ -41,21 +55,31 @@ impl core::fmt::Debug for SbiRet {
     }
 }
 
+/// RISC-V SBI error in enumeration
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Error {
+    /// Error for SBI call failed for unknown reasons
     Failed,
+    /// Error for target operation not supported
     NotSupported,
+    /// Error for invalid parameter
     InvalidParam,
+    /// Error for denied (unused in standard extensions)
     Denied,
+    /// Error for invalid address
     InvalidAddress,
+    /// Error for resource already available
     AlreadyAvailable,
+    /// Error for resource already started
     AlreadyStarted,
+    /// Error for resource already stopped
     AlreadyStopped,
+    /// Custom error code
     Custom(isize),
 }
 
 impl SbiRet {
-    /// Returns success SBi state with given `value`.
+    /// Returns success SBI state with given `value`.
     #[inline]
     pub const fn ok(value: usize) -> Self {
         Self {
@@ -156,13 +180,13 @@ impl SbiRet {
         }
     }
 
-    /// Returns `true` if current SBI return succeeded
+    /// Returns `true` if current SBI return succeeded.
     #[inline]
     pub const fn is_ok(&self) -> bool {
         matches!(self.error, RET_SUCCESS)
     }
 
-    /// Returns `true` if current SBI return is an error
+    /// Returns `true` if current SBI return is an error.
     #[inline]
     pub const fn is_err(&self) -> bool {
         !self.is_ok()