+++ src/hsm.rs
@@ -1,5 +1,6 @@
 //! Chapter 9. Hart State Management Extension (EID #0x48534D "HSM")
 
+/// Extension ID for Hart State Management extension
 pub const EID_HSM: usize = crate::eid_from_str("HSM") as _;
 pub use fid::*;
 
@@ -20,7 +21,7 @@ pub const HART_STATE_START_PENDING: usize = 2;
 /// The hart has requested to stop (or power-down) itself from the STARTED state
 /// and the SBI implementation is still working to get the hart in the STOPPED state.
 pub const HART_STATE_STOP_PENDING: usize = 3;
-/// This hart is in a platform specific suspend (or low power) state
+/// The hart is in a platform specific suspend (or low power) state
 pub const HART_STATE_SUSPENDED: usize = 4;
 /// The hart is pending before being suspended
 ///
@@ -40,14 +41,22 @@ pub const HART_SUSPEND_TYPE_RETENTIVE: u32 = 0;
 /// Default non-retentive hart suspend type
 pub const HART_SUSPEND_TYPE_NON_RETENTIVE: u32 = 0x8000_0000;
 
-/// §9.5
+/// Declared in §9.5
 mod fid {
-    /// §9.1
+    /// Function ID to start executing the given hart at specified address in supervisor-mode
+    ///
+    /// Declared in §9.1
     pub const HART_START: usize = 0;
-    /// §9.2
+    /// Function ID to stop executing the calling hart in supervisor-mode
+    ///
+    /// Declared in §9.2
     pub const HART_STOP: usize = 1;
-    /// §9.3
+    /// Function ID to get the current status (or HSM state id) of the given hart
+    ///
+    /// Declared in §9.3
     pub const HART_GET_STATUS: usize = 2;
-    /// §9.4
+    /// Function ID to put the calling hart into suspend or platform specific lower power states
+    ///
+    /// Declared in §9.4
     pub const HART_SUSPEND: usize = 3;
 }