+++ CHANGELOG.md
@@ -9,8 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Function `SbiRet::is_ok`, `is_err` and `expect`
+- Various documents
 
 ### Modified
+
+- Rename `SbiSpecVersion` to struct `Version` 
 
 ### Removed