+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Function `SbiRet::is_ok`, `is_err` and `expect`
-- Various documents
+- Add documents on whole `sbi-rt` crate
 
 ### Modified