+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ### Added
 
 - Function `SbiRet::is_ok`, `is_err` and `expect`
-- Add documents on whole `sbi-rt` crate
+- Add documents on whole `sbi-rt` crate to coply with `deny(missing_docs)`
+- Feature `legacy` to gate legacy SBI extension
 
 ### Modified