+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+### Modified
+
+### Fixed
+
+## [0.0.4] - 2022-10-10
+
+### Added
+
 - Various convenient functions to `SbiRet` structure
 - Add documents on whole `sbi-rt` crate to coply with `deny(missing_docs)`
 - Feature `legacy` to gate legacy SBI extension
@@ -16,8 +24,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - Rename `SbiRet::ok` to `SbiRet::success`
 - Rename `SbiSpecVersion` to struct `Version` 
-
-### Removed
 
 ## [0.0.3] - 2022-10-06
 
@@ -56,7 +62,8 @@ This is the first release of sbi-spec crate. This crate includes definition of R
 
 - Adapt to SBI specification version 1.0.0 ratified
 
-[Unreleased]: https://github.com/rustsbi/sbi-spec/compare/v0.0.3...HEAD
+[Unreleased]: https://github.com/rustsbi/sbi-spec/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/rustsbi/sbi-spec/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/rustsbi/sbi-spec/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/rustsbi/sbi-spec/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/rustsbi/sbi-spec/releases/tag/v0.0.1